addvisits visits details

+++ hospital/src/main/java/hospital/hospital/doctor/models/DoctorREQ.java
@@ -14,5 +14,5 @@ public class DoctorREQ extends UserDTO {
     private String name;
     private String surname;
     private Set<Long> specialisations;
-    private Long room;
+    private Integer room;
 }
